clean and split the App file
